// src/services/clients/clients.d.ts
export interface IProductsClient extends Service {
  getProducts(
    type: EProductTypes,
    location: string,
    query: string, 
    page?: number,
    pageSize?: number,
  ): Promise<IJsonResponse>;

  getProduct(id: number): Promise<IJsonResponse>;

  getProductsByOrganization(
    id: number,
    type: string, 
    query: string,
    page?: number, 
    pageSize?: number
  ): Promise<IJsonResponse>;

  createProduct(
    name: string,
    description: string,
    image: File,
    price: number,
    quantityInStock: number,
    dosage: string,
    drugType: EDrugTypes,
    productType: EProductTypes
  ): Promise<IJsonResponse>;

  createProductForOrganization(
    organizationId: number,
    name: string,
    description: string,
    image: File,
    price: number,
    quantityInStock: number,
    dosage: string,
    drugType: EDrugTypes,
    productType: EProductTypes 
  ): Promise<IJsonResponse>;

  deleteProduct(id: number): Promise<IJsonResponse>;

  updateProduct(
    id: number,
    name: string,
    description: string,
    price: number,
    quantityInStock: number,
    dosage: string,
    drugType: string,
    productType: string
  ): Promise<IJsonResponse>;

  changeProductImage(id: number, image: File): Promise<IJsonResponse>;
}